+++ LitterBox/LBEngine/src/LitterBox/Serialization/AssetManager.cpp
@@ -894,8 +894,8 @@ namespace LB
                 };
                 Characters.emplace(std::pair<char, Renderer::Character>(c, sc));
 
-                maxAscent = maxAscent < font->glyph->bitmap_top ? font->glyph->bitmap_top : maxAscent;
-                maxDescent = maxDescent < font->glyph->metrics.height - font->glyph->bitmap_top ?
+                maxAscent = static_cast<int>(maxAscent) < font->glyph->bitmap_top ? font->glyph->bitmap_top : maxAscent;
+                maxDescent = static_cast<int>(maxDescent) < font->glyph->metrics.height - font->glyph->bitmap_top ?
                     font->glyph->metrics.height - font->glyph->bitmap_top : maxDescent;
             }
             //insert the height
@@ -947,8 +947,8 @@ namespace LB
                 };
                 Characters.emplace(std::pair<char, Renderer::Character>(c, sc));
 
-                maxAscent = maxAscent < font->glyph->bitmap_top ? font->glyph->bitmap_top : maxAscent;
-                maxDescent = maxDescent < font->glyph->metrics.height - font->glyph->bitmap_top ?
+                maxAscent = static_cast<int>(maxAscent) < font->glyph->bitmap_top ? font->glyph->bitmap_top : maxAscent;
+                maxDescent = static_cast<int>(maxDescent) < font->glyph->metrics.height - font->glyph->bitmap_top ?
                     font->glyph->metrics.height - font->glyph->bitmap_top : maxDescent;
             }
             //insert the height
